+++ moderngl/texture.py
@@ -1,10 +1,14 @@
+import logging
 from typing import Tuple
 
+from moderngl.mgl import InvalidObject  # type: ignore
 from .buffer import Buffer
 
 __all__ = ['Texture',
            'NEAREST', 'LINEAR', 'NEAREST_MIPMAP_NEAREST', 'LINEAR_MIPMAP_NEAREST', 'NEAREST_MIPMAP_LINEAR',
            'LINEAR_MIPMAP_LINEAR']
+
+LOG = logging.getLogger(__name__)
 
 #: Returns the value of the texture element that is nearest 
 #: (in Manhattan distance) to the specified texture coordinates. 
@@ -66,13 +70,21 @@ class Texture:
         raise TypeError()
 
     def __repr__(self):
-        return '<Texture: %d>' % self.glo
+        if hasattr(self, '_glo'):
+            return f"<{self.__class__.__name__}: {self._glo}>"
+        else:
+            return f"<{self.__class__.__name__}: INCOMPLETE>"
 
     def __eq__(self, other):
         return type(self) is type(other) and self.mglo is other.mglo
 
     def __hash__(self) -> int:
         return id(self)
+
+    def __del__(self):
+        LOG.debug(f"{self.__class__.__name__}.__del__ {self}")
+        if hasattr(self, "ctx") and self.ctx.gc_mode == "auto":
+            self.release()
 
     @property
     def repeat_x(self) -> bool:
@@ -130,6 +142,10 @@ class Texture:
 
         return self.mglo.filter
 
+    @filter.setter
+    def filter(self, value):
+        self.mglo.filter = value
+
     @property
     def anisotropy(self) -> float:
         '''
@@ -149,10 +165,6 @@ class Texture:
     @anisotropy.setter
     def anisotropy(self, value):
         self.mglo.anisotropy = value
-
-    @filter.setter
-    def filter(self, value):
-        self.mglo.filter = value
 
     @property
     def swizzle(self) -> str:
@@ -298,6 +310,11 @@ class Texture:
         '''
             Read the pixel data as bytes into system memory.
 
+            The texture can also be attached to a :py:class:`Framebuffer`
+            to gain access to :py:meth:`Framebuffer.read` for additional
+            features such ad reading a subsection or converting to
+            another ``dtype``.
+
             Keyword Args:
                 level (int): The mipmap level.
                 alignment (int): The byte alignment of the pixels.
@@ -353,9 +370,16 @@ class Texture:
                 texture = ctx.texture((2, 2), 1)
                 texture.write(data)
 
+                # Write to a sub-section of the texture using viewport
+                texture = ctx.texture((100, 100), 4)
+                # Fill the lower left 50x50 pixels with new data
+                texture.write(data, viewport=(0, 0, 50, 50))
+
             Args:
                 data (Union[bytes, Buffer]): The pixel data.
-                viewport (tuple): The viewport.
+                viewport (tuple): The sub-section of the texture to update
+                                  in viewport coordinates. The data size
+                                  must match the size of the area.
 
             Keyword Args:
                 level (int): The mipmap level.
@@ -446,5 +470,6 @@ class Texture:
         '''
             Release the ModernGL object.
         '''
-
-        self.mglo.release()
+        LOG.debug(f"{self.__class__.__name__}.release() {self}")
+        if not isinstance(self.mglo, InvalidObject):
+            self.mglo.release()