+++ tests/test_transform.py
@@ -1,4 +1,5 @@
 import struct
+import platform
 from array import array
 import unittest
 import numpy as np
@@ -33,6 +34,7 @@ class TestCase(unittest.TestCase):
 
         self.assertIs(program.geometry_input, None)
         self.assertIs(program.geometry_output, None)
+        self.assertIs(program.is_transform, True)
 
         self.assertIn('vert', program)
         self.assertIn('vert_length', program)
@@ -56,6 +58,7 @@ class TestCase(unittest.TestCase):
         )
 
         vao = self.ctx.vertex_array(program, [(buffer, "2f", "in_pos"),])
+        self.assertIs(vao.mode, moderngl.POINTS)
         buffer2 = self.ctx.buffer(reserve=buffer.size)
         vao.transform(buffer2, mode=self.ctx.POINTS)
         self.assertEqual(data, struct.unpack(f"{len(data)}f", buffer.read()))
@@ -88,6 +91,9 @@ class TestCase(unittest.TestCase):
         self.assertEqual(tuple(v * 2 for v in data), struct.unpack(f"{len(data)}f", buffer2.read()))
 
     def test_vertex_lines_adjacency(self):
+        if platform.system().lower() in ["darwin"]:
+            self.skipTest('Tranforms with adjacency primitives now working on OSX')
+
         data = (
             1.0, 1.0,
             2.0, 2.0,
@@ -145,6 +151,9 @@ class TestCase(unittest.TestCase):
         )
 
     def test_vertex_line_strip_adjacency(self):
+        if platform.system().lower() in ["darwin"]:
+            self.skipTest('Tranforms with adjacency primitives now working on OSX')
+
         data = (
             1.0, 1.0,
             2.0, 2.0,
@@ -206,6 +215,8 @@ class TestCase(unittest.TestCase):
         )
 
     def test_vertex_triangles_adjacency(self):
+        if platform.system().lower() in ["darwin"]:
+            self.skipTest('Tranforms with adjacency primitives now working on OSX')
         data = (
             1.0, 1.0, # outer points
             2.0, 2.0,