+++ tests/test_vertex_array.py
@@ -1,4 +1,5 @@
 from array import array
+from moderngl.program_members import varying
 import unittest
 
 import moderngl
@@ -24,6 +25,7 @@ class TestCase(unittest.TestCase):
                 out_pos = pos + velocity;
             }
             """,
+            varyings=["out_pos"],
         )
         buffer = self.ctx.buffer(array('f', range(16)))
         self.ctx.vertex_array(prog, [(buffer, '2f 2x4', 'pos')])
@@ -75,6 +77,7 @@ class TestCase(unittest.TestCase):
         vbo1 = self.ctx.buffer(np.array([4.0, 2.0, 7.5, 1.8], dtype='f4').tobytes())
         vbo2 = self.ctx.buffer(reserve=vbo1.size)
         vao = self.ctx.simple_vertex_array(prog, vbo1, 'in_vert')
+        self.assertEqual(vao.mode, moderngl.POINTS)
         vao.transform(vbo2, moderngl.POINTS)
         res = np.frombuffer(vbo2.read(), dtype='f4')
         np.testing.assert_almost_equal(res, [4.0, 2.0, 7.5, 1.8])