+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+# 5.7.0
+
+* Integer textures is not using `NEAREST` interpolation by default
+* Added `Program.is_transform` exposing if the program has a fragment shader or not
+* Added `VertexArray.mode` and a `mode` argument in `Context.vertex_array`.
+  This is now the default rendering mode when no mode parameter is passed
+  in `render()` or `transform()`
+* Added support for 1D sampler in `Uniform`
+* Added direct access to `glEnable` / `glDisable` in `Context.enable_direct` / `Context.disable_direct`.
+  This can be used to enabled capabilities not supported by ModernGL.
+* Docstring improvements
+* Documentation improvements
+
 # [5.6.4]
 
 * Transform feedback now supports all primitive modes both for